创建MQ数据开关
